Rename keep alive publisher to heartbeat publisher
